+++ src/structures/dataSet.cpp
@@ -19,7 +19,7 @@ void DataSet::load() {
 }
 
 void DataSet::showOrderBy(const std::vector<int>& sorted_indices) {
-  for (int index : sorted_indices) {
+  for (size_t index : sorted_indices) {
     if (index >= 0 && index < data.size()) {
       std::cout << data[index] << " ";
     }
